Directly consume options from Ec2ServerCreate & Bootstrap plugins.
This should help to bring in all relavent options from these 2 plugins
rather than having to duplicate the configuration in multiple places.

The tricky implementation details of this are in the `deps` block. Since
this block is called when the plugin's #load_deps method is called any
class-defined options will have already been added to the `self.options`
Hash. This is the high level idea:

1. Options from Knife::ServerBootstrapBase will be loaded via the
  `include` line.
2. Local options will be loaded next, possibly overriding values from
   Knife::ServerBootstrapBase.
3. At a later time `#load_deps` is called on the object which will:
   1. Save the current options Hash in a variable for the moment.
   2. Load the options Hash from Knife::Bootstrap (or
      Knife::Ec2ServerCreate) replacing the object's options Hash.
   3. Merge back to original options Hash ontop of the current options
      Hash so that the options from Knife::Bootstrap have the lowest
      priority over common and local options that have already been
      defined.

In this way, the priority of options being loaded from lowest to highest
are:

1. Underlying Knife plugin such as Knife::Bootstrap
2. Common options from Knife::ServerBootstrapBase
3. Plugin-specific options from the class itself