[Breaking] Make knife-ec2 an optional gem dependency.
This was done to keep the dependencies of this gem lighter and to make
future adapters easier to add (using the same mechanism of lazy
loading).

Listing the Knife subcommands will still behave as normal and `knife
server boostrap ec2` will still appear. However if the knife-ec2 is not
present when a user attempts to use the subcommand (or display its full
help), a warning message will be displayed noting this requirement and
the subcommand will terminate with a non zero exit code.