Remove --chef-server-version in favor of --bootstrap-version.
The flag default value (which is nil) will evaluate to "11", and
combined with the updated default --platform value of "omnibus" will
result in a computed default value of "chef11/omnibus".

The major version number is determined by the value of
bootstrap_version, so any semver string (i.e. "10.24.0", "10", "10.18")
will result in a chef10/* template being chosen.

For Omnibus packages, the value of --bootstrap-version will be passed to
the package installer allowing the user to specify the desired version
of erchef (i.e. "11.4.0"). The --prerelease flag is also supported for
the chef11/omnibus template.